Added ModuleInterfaceRGBSwitch example
Added a switch to connect buses of the same or different strategies/physical media, based on the similar example in PJON, but implemented as a ModuleInterface module that reports status, uptime and traffic statistics.
This is great when having the master on a PC/RPI and devices running on Arduinos with SWBB or an other PJON strategy, using Ethernet to connect the master to the switches that connect to the different buses, and having a means to inspect and visualize the network and traffic.