Added metrics to switch to report module response times
This allows response time for packets throught the switch to be logged along with the responsible device id. It makes it possible to identify modules that have long unresponsive intervals, or devices that need a SWBB pulldown to improve the communication reliability. Plotting the reponse times gives much information about the bus and modules.